I've removed the pre-requirement that you need to have the 'bc' module installed.
It seems a little silly to have a hard-coded dependency on a module
that simply installs one package. I replaced it with a test to see if
the package has been defined elsewhere, and if not then it installs
it, allowing those who wish to continue to do this to have no
change.

I also provide a mechanism to specify the version, if desired.

Conflicts:

	manifests/linux.pp